Move EventEffects inline in events/handlers.ts
